packages/backend/requirements/pip.txt
bidict==0.23.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psutil==5.9.8
pycparser==2.21
pyOpenSSL==24.0.0
pyserial==3.5
pyzmq==25.1.2
simple-websocket==1.0.0
tornado==6.4
wsproto==1.2.0
opencv-python-headless==4.5.5.62